# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AirQr catalog and mock eSIM checkout API
 * OpenAPI spec version: 0.1.0
 */
export type ContactMessageInputSubject = typeof ContactMessageInputSubject[keyof typeof ContactMessageInputSubject];


export const ContactMessageInputSubject = {
  general: 'general',
  order_issue: 'order_issue',
  refund: 'refund',
  partnership: 'partnership',
} as const;

export interface ContactMessageInput {
  email: string;
  name: string;
  subject: ContactMessageInputSubject;
  message: string;
}

export type ContactMessageCreatedStatus = typeof ContactMessageCreatedStatus[keyof typeof ContactMessageCreatedStatus];


export const ContactMessageCreatedStatus = {
  new: 'new',
  resolved: 'resolved',
} as const;

export interface ContactMessageCreated {
  id: string;
  status: ContactMessageCreatedStatus;
  createdAt: string;
}

export type EsimStatusResultState = typeof EsimStatusResultState[keyof typeof EsimStatusResultState];


export const EsimStatusResultState = {
  active: 'active',
  inactive: 'inactive',
  exhausted: 'exhausted',
} as const;

export interface EsimStatusResult {
  state: EsimStatusResultState;
  dataRemainingGb: number;
  validityRemainingDays: number;
  lastSeenAt?: string;
}

export interface CountryStats {
  slug: string;
  ordersThisMonth: number;
}

export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Country {
  id: string;
  slug: string;
  name: string;
  flagEmoji: string;
  region: string;
  minPriceCents: number;
  tips?: string | null;
  planCount: number;
}

export interface Plan {
  id: string;
  countryId: string;
  countrySlug: string;
  countryName: string;
  flagEmoji: string;
  dataGb: number;
  validityDays: number;
  speed: string;
  coverage: string;
  networkOperator: string;
  hotspotAllowed: boolean;
  sellingPriceCents: number;
  currency: string;
}

export interface CountryDetail {
  country: Country;
  plans: Plan[];
}

export interface OrderInput {
  planId: string;
  email: string;
  name?: string;
  customerCountry?: string;
}

export interface OrderCreated {
  orderId: string;
  publicId: string;
  accessToken: string;
  totalCents: number;
  currency: string;
  status: string;
}

export interface PaymentConfirmation {
  orderId: string;
  accessToken: string;
}

export interface PaymentConfirmationResult {
  publicId: string;
  accessToken: string;
  status: string;
  paymentStatus: string;
  esimStatus: string;
}

export interface AccountLinkInput {
  email: string;
}

export interface AccountLinkResult {
  message: string;
  mockLink?: string;
}

export interface AccountOrder {
  publicId: string;
  accessToken: string;
  email: string;
  countryName: string;
  flagEmoji: string;
  dataGb: number;
  validityDays: number;
  totalCents: number;
  currency: string;
  status: string;
  esimStatus: string;
  createdAt: string;
  paymentStatus: string;
}

export interface Order {
  publicId: string;
  email: string;
  /** @nullable */
  name: string | null;
  countryName: string;
  flagEmoji: string;
  dataGb: number;
  validityDays: number;
  speed: string;
  coverage: string;
  networkOperator: string;
  hotspotAllowed: boolean;
  totalCents: number;
  currency: string;
  status: string;
  paymentStatus: string;
  esimStatus: string;
  createdAt: string;
  /** @nullable */
  qrPayload: string | null;
  /** @nullable */
  activationCode: string | null;
  /** @nullable */
  smdpAddress: string | null;
  /** @nullable */
  matchingId: string | null;
}

export type WebhookInputPayload = { [key: string]: unknown };

export interface WebhookInput {
  eventId: string;
  provider: string;
  payload: WebhookInputPayload;
}

export interface WebhookResult {
  accepted: boolean;
  duplicate: boolean;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

/**
 * Account session required
 */
export type UnauthorizedResponse = Error;

export type RegionParameter = string;

export type SearchParameter = string;

export type MaxPriceCentsParameter = number;

/**
 * Order access token that proves ownership
 */
export type OrderAccessTokenParameter = string;

export type ListCountriesParams = {
region?: RegionParameter;
search?: SearchParameter;
/**
 * @minimum 0
 */
maxPriceCents?: MaxPriceCentsParameter;
};

export type GetOrderParams = {
token?: string;
};

export type GetEsimStatusParams = {
publicId: string;
/**
 * Order access token that proves ownership
 */
token: OrderAccessTokenParameter;
};