# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AirQr catalog and mock eSIM checkout API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a pending order for one eSIM plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Pending order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
  /orders/{publicId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a protected order and eSIM delivery data
      parameters:
        - $ref: "#/components/parameters/PublicId"
        - name: token
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          description: Access token required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    PublicId:
      name: publicId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    OrderInput:
      type: object
      properties:
        planId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        customerCountry:
          type: string
      required: [planId, email]
    OrderCreated:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        publicId:
          type: string
          format: uuid
        accessToken:
          type: string
          format: uuid
        totalCents:
          type: integer
        currency:
          type: string
        status:
          type: string
      required: [orderId, publicId, accessToken, totalCents, currency, status]
...
    Order:
      type: object
      properties:
        publicId:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: ["string", "null"]
        countryName:
          type: string
        flagEmoji:
          type: string
        dataGb:
          type: integer
        validityDays:
          type: integer
        speed:
          type: string
        coverage:
          type: string
        networkOperator:
          type: string
        hotspotAllowed:
          type: boolean
        totalCents:
          type: integer
        currency:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        esimStatus:
          type: string
        createdAt:
          type: string
          format: date-time
        qrPayload:
          type: ["string", "null"]
        activationCode:
          type: ["string", "null"]
        smdpAddress:
          type: ["string", "null"]
        matchingId:
          type: ["string", "null"]
      required:
        - publicId
        - email
        - name
        - countryName
        - flagEmoji
        - dataGb
        - validityDays
        - speed
        - coverage
        - networkOperator
        - hotspotAllowed
        - totalCents
        - currency
        - status
        - paymentStatus
        - esimStatus
        - createdAt
        - qrPayload
        - activationCode
        - smdpAddress
        - matchingId
...