# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AirQr catalog and mock eSIM checkout API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  publicId: string;
  email: string;
  /** @nullable */
  name: string | null;
  countryName: string;
  flagEmoji: string;
  dataGb: number;
  validityDays: number;
  speed: string;
  coverage: string;
  networkOperator: string;
  hotspotAllowed: boolean;
  totalCents: number;
  currency: string;
  status: string;
  paymentStatus: string;
  esimStatus: string;
  createdAt: Date;
  /** @nullable */
  qrPayload: string | null;
  /** @nullable */
  activationCode: string | null;
  /** @nullable */
  smdpAddress: string | null;
  /** @nullable */
  matchingId: string | null;
}